+++ show-viewshed-from-point-in-scene/src/main/java/com/esri/arcgismaps/sample/showviewshedfrompointinscene/screens/MainScreen.kt
@@ -61,219 +61,10 @@ fun MainScreen(sampleName: String, application: Application) {
                         modifier = Modifier.fillMaxSize().weight(1f),
                         sceneViewModel = sceneViewModel
                     )
-                    // sliders
-                    HeadingSlider(sceneViewModel)
-                    PitchSlider(sceneViewModel)
-                    HorizontalAngleSlider(sceneViewModel)
-                    VerticalAngleSlider(sceneViewModel)
-                    MinimumDistanceSlider(sceneViewModel)
-                    MaximumDistanceSlider(sceneViewModel)
-                    // checkbox
-                    Row {
-                        FrustumCheckBox(sceneViewModel)
-                        AnalysisCheckBox(sceneViewModel)
-                    }
+                    ViewshedOptionsScreen(sceneViewModel)
                 }
 
             }
         }
     )
-}
-
-@Composable
-private fun HeadingSlider(sceneViewModel: SceneViewModel) {
-
-    var sliderValue by remember {
-        mutableStateOf(82f)
-    }
-    Row {
-        Text(modifier = Modifier.padding(start = 10.dp, top = 10.dp, end = 10.dp).width(150.dp),
-            text = "Heading")
-        Slider(
-            modifier = Modifier.weight(1f),
-            value = sliderValue,
-            onValueChange = {
-                sliderValue = it
-                // this is called when the user completed selecting the value
-                sceneViewModel.setHeading(sliderValue)
-            },
-            valueRange = 0f..360f
-        )
-        Text(
-            modifier = Modifier.padding(start = 10.dp, top = 10.dp, end = 10.dp).size(40.dp),
-            text = sliderValue.toInt().toString()
-        )
-    }
-}
-
-@Composable
-private fun PitchSlider(sceneViewModel: SceneViewModel) {
-
-    var sliderValue by remember {
-        mutableStateOf(60f)
-    }
-    Row {
-        Text(modifier = Modifier.padding(start = 10.dp, end = 10.dp, top = 10.dp).width(150.dp),
-            text = "Pitch")
-        Slider(
-            modifier = Modifier.weight(1f),
-            value = sliderValue,
-            onValueChange = {
-                sliderValue = it
-                // this is called when the user completed selecting the value
-                sceneViewModel.setPitch(sliderValue)
-            },
-            valueRange = 0f..180f
-        )
-        Text(
-            modifier = Modifier.padding(start = 10.dp, end = 10.dp, top = 10.dp).size(40.dp),
-            text = sliderValue.toInt().toString()
-        )
-    }
-}
-
-@Composable
-private fun HorizontalAngleSlider(sceneViewModel: SceneViewModel) {
-
-    var sliderValue by remember {
-        mutableStateOf(75f)
-    }
-    Row {
-        Text(
-            modifier = Modifier.padding(start = 10.dp, top = 10.dp, end = 10.dp).width(150.dp),
-            text = "Horizontal Angle"
-        )
-        Slider(
-            modifier = Modifier.weight(1f),
-            value = sliderValue,
-            onValueChange = {
-                sliderValue = it
-                // this is called when the user completed selecting the value
-                sceneViewModel.setHorizontalAngleSlider(sliderValue)
-            },
-            valueRange = 0f..120f
-        )
-        Text(
-            modifier = Modifier.padding(start = 10.dp, top = 10.dp, end = 10.dp).size(40.dp),
-            text = sliderValue.toInt().toString()
-        )
-    }
-}
-
-@Composable
-private fun VerticalAngleSlider(sceneViewModel: SceneViewModel) {
-
-    var sliderValue by remember {
-        mutableStateOf(90f)
-    }
-    Row {
-        Text(
-            modifier = Modifier.padding(start = 10.dp, top = 10.dp, end = 10.dp).width(150.dp),
-            text = "Vertical Angle"
-        )
-        Slider(
-            modifier = Modifier.weight(1f),
-            value = sliderValue,
-            onValueChange = {
-                sliderValue = it
-                // this is called when the user completed selecting the value
-                sceneViewModel.setVerticalAngleSlider(sliderValue)
-            },
-            valueRange = 0f..120f
-        )
-        Text(
-            modifier = Modifier.padding(start = 10.dp, top = 10.dp, end = 10.dp).size(40.dp),
-            text = sliderValue.toInt().toString()
-        )
-    }
-}
-
-@Composable
-private fun MinimumDistanceSlider(sceneViewModel: SceneViewModel) {
-
-    var sliderValue by remember {
-        mutableStateOf(0f)
-    }
-    Row {
-        Text(
-            modifier = Modifier.padding(start = 10.dp, top = 10.dp, end = 10.dp).width(150.dp),
-            text = "Minimum Distance"
-        )
-        Slider(
-            modifier = Modifier.weight(1f),
-            value = sliderValue,
-            onValueChange = {
-                sliderValue = it
-                // this is called when the user completed selecting the value
-                sceneViewModel.setMinimumDistanceSlider(sliderValue)
-            },
-            valueRange = 0f..8999f
-        )
-        Text(
-            modifier = Modifier.padding(start = 10.dp, top = 10.dp, end = 10.dp).size(40.dp),
-            text = sliderValue.toInt().toString()
-        )
-    }
-}
-
-@Composable
-private fun MaximumDistanceSlider(sceneViewModel: SceneViewModel) {
-
-    var sliderValue by remember {
-        mutableStateOf(1500f)
-    }
-    Row {
-        Text(
-            modifier = Modifier.padding(start = 10.dp, top = 10.dp, end = 10.dp).width(150.dp),
-            text = "Maximum Distance"
-        )
-        Slider(
-            modifier = Modifier.weight(1f),
-            value = sliderValue,
-            onValueChange = {
-                sliderValue = it
-                // this is called when the user completed selecting the value
-                sceneViewModel.setMaximumDistanceSlider(sliderValue)
-            },
-            valueRange = 0f..9999f
-        )
-        Text(
-            modifier = Modifier.padding(start = 10.dp, top = 10.dp, end = 10.dp).size(40.dp),
-            text = sliderValue.toInt().toString()
-        )
-    }
-}
-
-@Composable
-fun FrustumCheckBox(sceneViewModel: SceneViewModel) {
-    // set the state of the checkbox
-    val checkedState = remember { mutableStateOf(true) }
-    // display a row and create a checkbox and text in a row
-    Row {
-        Checkbox(
-            checked = checkedState.value,
-            onCheckedChange = {
-                checkedState.value = it
-                sceneViewModel.frustumVisibility(checkedState.value)
-            },
-        )
-        Text(modifier = Modifier.padding(top = 10.dp), text = "Frustum Outline")
-    }
-}
-
-@Composable
-fun AnalysisCheckBox(sceneViewModel: SceneViewModel) {
-    // set the state of the checkbox
-    val checkedState = remember { mutableStateOf(true) }
-    // display a row and create a checkbox and text in a row
-    Row {
-        Checkbox(
-            checked = checkedState.value,
-            onCheckedChange = {
-                checkedState.value = it
-                sceneViewModel.analysisVisibility(checkedState.value)
-            },
-        )
-        Text(modifier = Modifier.padding(top = 10.dp), text = "Analysis Overlay")
-    }
 }
